pyproject.toml
[project]
name = "learn-on-youtube-multi-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "langchain-google-genai>=2.1.9",
    "langchain-openai>=0.3.30",
    "langgraph-supervisor>=0.0.29",
]
